stage1/requirements.txt
pip>=24.0
pandas>=2.2
pyarrow>=16.0
python-dateutil>=2.9
pytz>=2024.1
PyYAML>=6.0
jsonschema>=4.22
python-dotenv>=1.0.1
requests>=2.32
